experiments/adr46-scanner/pyproject.toml
[project]
name = "adr46-scanner"
version = "0.1.0"
description = "Detect Tekton tasks drifting from ADR-0046 (common task runner image)"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
adr46-scan = "scanner.cli:main"
